Condense name and phone standardizations into a single module and update unit testing and python/IntakePipeline accordingly.
