+++ src/FunctionApps/python/IntakePipeline/__init__.py
@@ -20,8 +20,10 @@ from phdi_building_blocks.conversion import (
 )
 
 from phdi_building_blocks.geo import get_smartystreets_client, geocode_patient_address
-from phdi_building_blocks.name_standardization import standardize_patient_name
-from phdi_building_blocks.phone_standardization import standardize_patient_phone
+from phdi_building_blocks.standardize import (
+    standardize_patient_name,
+    standardize_patient_phone,
+)
 from phdi_building_blocks.linkage import add_patient_identifier
